Split types.hpp and field_collection.hpp in 2 headers each to isolate protobuf code.
