+++ cpp/arcticdb/entity/stream_descriptor.hpp
@@ -12,7 +12,7 @@
 #include <google/protobuf/util/message_differencer.h>
 #include <folly/gen/Base.h>
 
-#include <arcticdb/entity/field_collection.hpp>
+#include <arcticdb/entity/field_collection_proto.hpp>
 #include <arcticdb/util/variant.hpp>
 
 namespace arcticdb::entity {
